feat(service-log): extract a shared ServiceLogForm from AddForm
New presentation/ui/form/form.tsx + use-form.ts own useForm(), the
resolver, and the field JSX; AddForm/use-add.ts reduce to a thin
wrapper supplying only its mutation, mirroring Car's Add/Edit
precedent so EditForm can reuse the same fields and validation.

interface/ui/add-form.schema.ts renamed to service-log-form.schema.ts
(serviceLogFormSchema/ServiceLogFormValues) since it's no longer
add-specific.

Also defaults the date field to today instead of blank, computed
fresh per mount rather than at module load so a long-lived tab
doesn't hand out a stale date after midnight.
