+++ src/module/car/service-log/presentation/ui/forms/add/use-add.ts
@@ -1,49 +1,26 @@
-import { zodResolver } from '@hookform/resolvers/zod';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { useEffect } from 'react';
-import type { Resolver } from 'react-hook-form';
-import { useForm } from 'react-hook-form';
 
 import { serviceLogAddMutationOptions } from '@/car/service-log/infrastructure/tanstack/mutation-options/add';
-import type { AddServiceLogFormValues } from '@/car/service-log/interface/ui/add-form.schema';
-import { addServiceLogFormSchema } from '@/car/service-log/interface/ui/add-form.schema';
+import type { ServiceLogFormValues } from '@/car/service-log/interface/ui/service-log-form.schema';
 import { useToasts } from '@/common/presentation/hook/use-toasts';
 import { useSessionUser } from '@/user/presentation/hooks/use-session-user';
 
-import type { AddFormProps } from './add';
+interface UseAddFormParams {
+  carId: string;
+  onSubmit?: () => void;
+}
 
-const defaultAddFormValues: AddServiceLogFormValues = {
-  serviceDate: '',
-  categories: [],
-  mileage: null,
-  notes: null,
-  serviceCost: null,
-};
-
-export function useAddForm({ carId, onSubmit }: AddFormProps) {
+export function useAddForm({ carId, onSubmit }: UseAddFormParams) {
   const { data: sessionUser } = useSessionUser();
 
   const { addToast } = useToasts();
 
-  const {
-    register,
-    reset,
-    handleSubmit,
-    formState: { errors, isSubmitting, isValid, isDirty, isSubmitSuccessful },
-  } = useForm({
-    resolver: zodResolver(
-      addServiceLogFormSchema,
-    ) as Resolver<AddServiceLogFormValues>,
-    mode: 'onChange',
-    defaultValues: defaultAddFormValues,
-  });
-
   const queryClient = useQueryClient();
 
-  // onSubmit closes the add modal, unmounting this form before the request
-  // settles. React Query drops callbacks passed to mutate() once the caller
-  // unmounts, so the toasts live on the mutation options (which still run).
-  // onError is composed so the options' optimistic rollback still runs.
+  // The modal unmounts this form as soon as onSubmit fires, before the
+  // request settles. React Query drops callbacks passed to mutate() once the
+  // caller unmounts, so the toasts live on the mutation options (which still
+  // run). onError is composed so the options' optimistic rollback still runs.
   const addMutationOptions = serviceLogAddMutationOptions(queryClient);
 
   const { mutate } = useMutation({
@@ -55,24 +32,10 @@ export function useAddForm({ carId, onSubmit }: AddFormProps) {
     },
   });
 
-  useEffect(() => {
-    isSubmitSuccessful && reset();
-  }, [isSubmitSuccessful, reset]);
-
-  const handleFormSubmit = handleSubmit((formData: AddServiceLogFormValues) => {
+  const handleFormSubmit = (formData: ServiceLogFormValues) => {
     onSubmit && onSubmit();
     mutate({ carId, authorId: sessionUser?.id ?? '', ...formData });
-  });
-
-  const handleFormReset = () => reset();
-
-  return {
-    handleFormSubmit,
-    handleFormReset,
-    register,
-    errors,
-    isSubmitting,
-    isValid,
-    isDirty,
   };
+
+  return { handleFormSubmit };
 }
